evaluation/requirement.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
ansi2html==1.9.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
apptools==5.2.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.0
comm==0.2.1
configargparse==1.5.3
configobj==5.0.8
cprint==1.2.2
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
decorator==5.1.1
deprecated==1.2.14
docker-pycreds==0.4.0
easydict==1.11
einops==0.4.1
entrypoints==0.4
envisage==7.0.3
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flask==3.0.0
flatbuffers==24.3.25
fonttools==4.47.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2023.12.2
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
geometrout==0.1.3
gitdb==4.0.11
gitpython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.7.0
huggingface-hub==0.22.2
hydra-core==1.2.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
iopath==0.1.10
ipycanvas==0.13.1
ipyevents==2.0.2
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-client==7.4.9
jupyter-core==5.7.0
jupyterlab-widgets==3.0.9
keras==2.13.1
kiwisolver==1.4.5
lazy-loader==0.3
libclang==18.1.1
llvmlite==0.41.1
lmdb==1.4.1
loguru==0.6.0
lxml==5.0.0
markdown==3.5.1
markupsafe==2.1.3
marshmallow==3.21.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mayavi==4.8.2
mesh2sdf==1.1.0
meshcat==0.3.2
mpmath==1.3.0
msgpack==1.0.7
msgpack-numpy==0.4.8
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.2.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.6
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.0
oauthlib==3.2.2
omegaconf==2.2.3
open3d==0.18.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.2
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
plotly==5.11.0
pluggy==1.4.0
plyfile==1.0.2
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pybullet==3.2.6
pycollada==0.7.2
pyface==8.0.0
pyglet==1.5.27
pygltflib==1.16.2
pygments==2.17.2
pyngrok==7.0.5
pyopengl==3.1.0
pyparsing==3.1.1
pyqt5==5.15.10
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pyquaternion==0.9.9
pyrender==0.1.45
pyribbit==0.1.46
pytest==8.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywavelets==1.4.1
pyyaml==6.0.1
pyzmq==24.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.16.2
rsa==4.9
rtree==1.1.0
sacremoses==0.1.1
safetensors==0.4.2
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==1.39.1
setproctitle==1.3.3
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
svg-path==6.3
sympy==1.12
tabulate==0.8.10
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.10.3
tomli==2.0.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
traits==6.4.3
traitsui==8.0.0
transformations==2022.9.26
transformers==4.10.0
transforms3d==0.4.1
trimesh==3.23.1
triton==2.1.0
typing-extensions==4.5.0
typing-inspect==0.9.0
tzdata==2023.4
u-msgpack-python==2.8.0
urchin==0.0.27
urdf-parser-py==0.0.4
urllib3==2.1.0
usd-core==23.5
vtk==9.3.0
wandb==0.16.1
wcwidth==0.2.13
werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
yourdfpy==0.0.53
zipp==3.17.0